1、web模块新增参数校验以及全局枚举 2、新增搭建租户模块 3、安全框架中新增认证上下文功能 4、搭建mybatis-plus结构(50%) 5、在mybatis-flex原有框架基础上增加拦截器功能 6、实现基于mybatis-flex的数据权限功能，以及租户功能 7、增强mybatis-flex的table功能，支持判空操作 8、新增基于TypeHandler，处理数据库数组数据 9、新增mybatis-flex的插入更新感知，自动填充代码 10、接入工作流camunda（80%） 11、搭建nacos模块，目前暂未实现 12、在common模块新增工具类

+++ gloam-framework-web/src/main/java/com/gloamframework/web/response/WebPage.java
@@ -22,15 +22,15 @@ public class WebPage<T> extends Result<Collection<T>> {
     private static final int DEFAULT_TOTAL = 0;
 
     @ApiModelProperty(value = "分页页数")
-    private final int pageNum;
+    private final long pageNum;
 
     @ApiModelProperty(value = "分页大小")
-    private final int pageSize;
+    private final long pageSize;
 
     @ApiModelProperty(value = "分页总数")
-    private final int total;
+    private final long total;
 
-    private WebPage(Collection<T> data, HttpStatus status, boolean success, String message, int pageNum, int pageSize, int total, Object... params) {
+    private WebPage(Collection<T> data, HttpStatus status, boolean success, String message, long pageNum, long pageSize, long total, Object... params) {
         super(data, status, success, message, params);
         this.pageNum = pageNum;
         this.pageSize = pageSize;
@@ -43,11 +43,11 @@ public class WebPage<T> extends Result<Collection<T>> {
      * @param data    传输数据
      * @param message 传输消息
      */
-    public static <T> WebPage<T> success(Collection<T> data, int pageNum, int pageSize, int total, String message, Object... params) {
+    public static <T> WebPage<T> success(Collection<T> data, long pageNum, long pageSize, long total, String message, Object... params) {
         return new WebPage<>(data, HttpStatus.OK, true, message, pageNum, pageSize, total, params);
     }
 
-    public static <T> WebPage<T> success(Collection<T> data, int pageNum, int pageSize, int total) {
+    public static <T> WebPage<T> success(Collection<T> data, long pageNum, long pageSize, long total) {
         return success(data, pageNum, pageSize, total, null);
     }
